single-calendar changes
I am SO SORRY I took so long getting back to you, I am so embarrassed. I got pulled in a bunch of different directions.
Anyways, it looks like you have a solution to your issue but I went ahead and added another way that may or may not make more sense. Keep up the great work and feel free to reach out to us anytime in the AskBCS channel!
-Wilson